requirements.txt
Django==4.1.7
django-jalali==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ipython==8.11.0
jdatetime==4.1.1
Pygments==2.14.0
PyJWT==2.6.0
python-dotenv==1.0.0
python-env==1.0.0
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
djangorestframework-swagger==0.2.2
shortuuid==1.0.11
psycopg2-binary==2.9.6
python-dotenv==1.0.0
bcrypt==4.0.1